requirements.txt
asgiref==3.5.2
Django==4.1.3
django-searchable-encrypted-fields==0.2.1
djangorestframework==3.14.0
mysqlclient==2.1.1
numpy==1.21.6
pyasn1==0.4.8
pycrypto==2.6.1
pycryptodome==3.16.0
python-keyczar==0.716
pytz==2022.6
sqlparse==0.4.3
uWSGI==2.0.21
zipp==3.8.1
